requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
curl_cffi==0.11.4
frozendict==2.4.6
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
multitasking==0.0.11
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.0
peewee==3.18.1
platformdirs==4.3.8
pluggy==1.6.0
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.63
